# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PCMI API
  description: Persistent Cognitive Memory Infrastructure — memory API (v1)
  version: 1.51.0
servers:
  - url: http://localhost:8000
    description: Local
security:
  - ApiKeyAuth: []
paths:
...
  /v1/graph/related:
    get:
      summary: Traverse memory links via AGE (experimental spike)
      description: |
        Returns memories reachable within `depth` hops from `memory_id`.
        Requires AGE; returns 501 when unavailable. See [cognitive-graph.md](cognitive-graph.md).
      parameters:
        - name: memory_id
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: depth
          in: query
          schema: { type: integer, minimum: 1, maximum: 10, default: 3 }
        - name: link_types
          in: query
          schema: { type: string }
          description: Comma-separated link types (causal, temporal, contradicts, supports, related)
        - name: direction
          in: query
          schema: { type: string, enum: [both, out, in], default: both }
          description: Traversal direction — both follows edges in either direction; out is legacy outgoing-only
        - name: cursor
          in: query
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
          description: Keyset pagination cursor (last memory ID from previous page)
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Related memories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphRelatedResponse"
        "400":
          description: Invalid query
        "501":
          description: AGE not installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphUnavailableResponse"
  /v1/graph/chain:
    get:
      summary: Find shortest causal chain between two memories (experimental spike)
      description: |
        Finds the shortest path between two memories using the given link types.
        Requires AGE; returns 501 when unavailable. See [cognitive-graph.md](cognitive-graph.md).
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: integer, format: int64 }
          description: Source memory ID
        - name: to
          in: query
          required: true
          schema: { type: integer, format: int64 }
          description: Target memory ID
        - name: link_types
          in: query
          schema: { type: string }
          description: Comma-separated link types (causal, temporal, contradicts, supports, related)
        - name: max_depth
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 10 }
      responses:
        "200":
          description: Chain result (connected or not)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphChainResponse"
        "400":
          description: Invalid query
        "501":
          description: AGE not installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphUnavailableResponse"
  /v1/graph/entities/memory:
    get:
      summary: List entity vertices mentioned by a memory (Phase B)
      description: |
        Returns promoted `:Entity` vertices linked from the memory via `:mentions`.
        Requires AGE and prior successful extraction with `entity_promotion` rules.
        Returns 501 when AGE is unavailable. See [cognitive-graph-entities.md](cognitive-graph-entities.md).
      parameters:
        - name: memory_id
          in: query
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Entity mentions for the memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphEntitiesForMemoryResponse"
        "400":
          description: Invalid query
        "501":
          description: AGE not installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphUnavailableResponse"
  /v1/graph/entities/related:
    get:
      summary: Correlate memories via shared entity vertices (Phase B)
      description: |
        Returns memories that mention the same entity (`kind` + `key`) or, when
        `memory_id` is given, other memories sharing any promoted entity with that memory.
        Requires AGE. Returns 501 when unavailable.
      parameters:
        - name: kind
          in: query
          schema: { type: string }
          description: Entity kind (e.g. IPAddress) — required with key unless memory_id is set
        - name: key
          in: query
          schema: { type: string }
          description: Normalized entity key — required with kind unless memory_id is set
        - name: memory_id
          in: query
          schema: { type: integer, format: int64 }
          description: Find memories sharing entities with this memory
        - name: cursor
          in: query
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Correlated memories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphEntitiesRelatedResponse"
        "400":
          description: Invalid query
        "501":
          description: AGE not installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphUnavailableResponse"
...
  /v1/graph/cypher:
    post:
      summary: Execute a read-only Cypher query (experimental spike; requires write role)
      description: |
        Passthrough for raw Cypher read queries against pcmi_memory_graph.
        Only MATCH queries are allowed. Tenant scoping is the caller's responsibility.
        Requires write role. See [cognitive-graph.md](cognitive-graph.md).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CypherQueryRequest"
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CypherQueryResponse"
        "400":
          description: Invalid query or query rejected
        "403":
          description: Read-only role cannot execute Cypher
        "501":
          description: AGE not installed
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    GraphRelatedEntry:
      type: object
      properties:
        id: { type: integer, format: int64 }
        link_type: { type: string }
        depth: { type: integer }
    GraphRelatedResponse:
      type: object
      properties:
        memory_id: { type: integer, format: int64 }
        depth: { type: integer }
        direction: { type: string, enum: [both, out, in] }
        count: { type: integer }
        total: { type: integer, description: Total matching entries before pagination }
        next_cursor: { type: integer, format: int64 }
        limit: { type: integer }
        entries:
          type: array
          items: { $ref: "#/components/schemas/GraphRelatedEntry" }
    GraphUnavailableResponse:
      type: object
      properties:
        error: { type: string }
        hint: { type: string }
    GraphChainLink:
      type: object
      properties:
        from_id: { type: integer, format: int64 }
        to_id: { type: integer, format: int64 }
        link_type: { type: string }
        hop: { type: integer }
    GraphChainResponse:
      type: object
      properties:
        from_id: { type: integer, format: int64 }
        to_id: { type: integer, format: int64 }
        connected: { type: boolean }
        hops: { type: integer }
        path:
          type: array
          items: { $ref: "#/components/schemas/GraphChainLink" }
    GraphEntityMention:
      type: object
      properties:
        slot: { type: string }
        kind: { type: string }
        key: { type: string }
        confidence: { type: number, format: float }
    GraphEntitiesForMemoryResponse:
      type: object
      properties:
        memory_id: { type: integer, format: int64 }
        count: { type: integer }
        entities:
          type: array
          items: { $ref: "#/components/schemas/GraphEntityMention" }
    GraphEntityRelatedEntry:
      type: object
      properties:
        id: { type: integer, format: int64 }
        shared_kind: { type: string }
        shared_key: { type: string }
        slot: { type: string }
        confidence: { type: number, format: float }
    GraphEntitiesRelatedResponse:
      type: object
      properties:
        kind: { type: string }
        key: { type: string }
        memory_id: { type: string }
        count: { type: integer }
        total: { type: integer }
        next_cursor: { type: integer, format: int64 }
        limit: { type: integer }
        entries:
          type: array
          items: { $ref: "#/components/schemas/GraphEntityRelatedEntry" }
...
    CypherQueryRequest:
      type: object
      required: [query]
      properties:
        query: { type: string, description: Read-only Cypher MATCH query }
    CypherQueryResponse:
      type: object
      properties:
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
...